+++ lending/CodatLending/Models/Operations/DownloadAccountingInvoicePdfResponse.cs
@@ -8,8 +8,9 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatLending.Models.Operations
+namespace Codat.Lending.Models.Operations
 {
+    using Codat.Lending.Models.Shared;
     using System.Net.Http;
     using System;
     
@@ -27,6 +28,11 @@ namespace CodatLending.Models.Operations
         public byte[]? Data { get; set; }
 
         /// <summary>
+        /// Your API request was not properly authorized.
+        /// </summary>
+        public ErrorMessage? ErrorMessage { get; set; }
+
+        /// <summary>
         /// HTTP response status code for this operation
         /// </summary>
         public int StatusCode { get; set; } = default!;
@@ -34,6 +40,6 @@ namespace CodatLending.Models.Operations
         /// <summary>
         /// Raw HTTP response; suitable for custom response parsing
         /// </summary>
-        public HttpResponseMessage? RawResponse { get; set; }
+        public HttpResponseMessage RawResponse { get; set; } = default!;
     }
 }