+++ lending/CodatLending/Models/Operations/ListAccountingBankAccountsRequest.cs
@@ -8,9 +8,9 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatLending.Models.Operations
+namespace Codat.Lending.Models.Operations
 {
-    using CodatLending.Utils;
+    using Codat.Lending.Utils;
     
     public class ListAccountingBankAccountsRequest
     {