+++ lending/CodatLending/Models/Shared/AccountingDirectCost.cs
@@ -8,7 +8,7 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatLending.Models.Shared
+namespace Codat.Lending.Models.Shared
 {
     using Newtonsoft.Json;
     using System.Collections.Generic;
@@ -37,10 +37,10 @@ namespace CodatLending.Models.Shared
     {
 
         /// <summary>
-        /// The customer or supplier for the transfer, if available.
+        /// A customer or supplier associated with the direct cost.
         /// </summary>
         [JsonProperty("contactRef")]
-        public ContactRef? ContactRef { get; set; }
+        public AccountingDirectCostContactRef? ContactRef { get; set; }
 
         /// <summary>
         /// The currency data type in Codat is the <a href="https://en.wikipedia.org/wiki/ISO_4217">ISO 4217</a> currency code, e.g. _GBP_.<br/>